Push/Pull Block
Player is able to "pull" block when they press E. They end the pulling by pressing E again (really I am just reparenting the block.)

You can only pull when you press E, you cannot push, I had to freeze physics on the RigidBody2D to get this to work.

Now then, all something needs to be able to be moved around like this is be a RigidBody2D, have a Size Variable, and be in the Block Group. However this is a gamejam, scope creep can be real, so realistically this is the only moveable thing we have.